Pack moves

+++ src/game/display.rs
@@ -97,10 +97,6 @@ static CHARS: &[char] = &['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i'];
 
 impl fmt::Display for super::Move {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(
-            f,
-            "{}{}",
-            CHARS[self.board as usize], CHARS[self.square as usize]
-        )
+        write!(f, "{}{}", CHARS[self.board()], CHARS[self.square()])
     }
 }
